requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.5.7
charset-normalizer==3.1.0
ctransformers==0.2.0
filelock==3.12.0
fsspec==2023.5.0
huggingface-hub==0.14.1
idna==3.4
packaging==23.1
python-telegram-bot==13.15
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
tornado==6.1
tqdm==4.65.0
typing_extensions==4.6.1
tzlocal==5.0.1
urllib3==2.0.2